coud renderer done

+++ Demos/FromShaderDesigner1.5.9.4/FormShaderDesigner1594Demos/Renderers/Cloud.frag
@@ -11,7 +11,7 @@
 
 in float LightIntensity; 
 in vec3  MCposition;
-uniform int  TIME_FROM_INIT;
+uniform float TIME_FROM_INIT;
 uniform sampler3D Noise;
 vec3 Offset = vec3(0,0,0);
 uniform vec3 SkyColor;     // (0.0, 0.0, 0.8)
@@ -19,7 +19,7 @@ uniform vec3 CloudColor;   // (0.8, 0.8, 0.8)
 
 void main (void)
 {  
-  float offset = float(TIME_FROM_INIT) *0.0001;
+  float offset = TIME_FROM_INIT *0.0001;
   Offset = vec3(-offset,offset,offset); // uncomment this line for animation
   vec4  noisevec  = texture3D(Noise, MCposition + Offset);  
   float intensity = (noisevec[0] + noisevec[1] +
